finished user roles demo

+++ lecture-18-user-roles/demo-1-user-roles-finish/routes/apiv1.js
@@ -4,14 +4,18 @@ var router = express.Router();
 /* GET users listing. */
 router.get('/users/me', (req, res) => {
     let session = req.session
+
     if(session.isAuthenticated){
-      res.json({
-          status: "logged in",
-          name: session.account.name,
-          username: session.account.username
-      });
+        let roles = session.account.idTokenClaims.roles ? session.account.idTokenClaims.roles : []
+        let isModerator = roles.includes("Content.Moderator")
+        res.json({
+            status: "logged in",
+            name: session.account.name,
+            username: session.account.username,
+            isModerator: isModerator
+        });
     } else {
-      res.json({status: "not logged in (anonymous)"})
+        res.json({status: "not logged in (anonymous)"})
     }
 });
 
@@ -34,10 +38,18 @@ router.post('/comments', async (req, res) => {
 
 
 router.delete("/comments", async (req, res, next) =>{
+    let session = req.session
+
     if(req.session.isAuthenticated){
-        let commentId = req.body.commentId
-        await req.models.Comment.deleteOne({_id: commentId})
-        res.json({status: "success"})
+        let roles = session.account.idTokenClaims.roles ? session.account.idTokenClaims.roles : []
+        let isModerator = roles.includes("Content.Moderator")
+        if(isModerator){
+            let commentId = req.body.commentId
+            await req.models.Comment.deleteOne({_id: commentId})
+            res.json({status: "success"})
+        } else {
+            res.json({status: "Error: Not a moderator"})
+        }
     } else {
         res.json({status: "Error: Not logged in"})
     }
